+++ src/builder/resettable.rs
@@ -1,9 +1,13 @@
 // Unlike `impl Into<Option<T>>` or `Option<impl Into<T>>`, this isn't ambiguous for the `None`
 // case.
 
+use crate::builder::ArgAction;
 use crate::builder::OsStr;
 use crate::builder::Str;
 use crate::builder::StyledStr;
+use crate::builder::ValueHint;
+use crate::builder::ValueParser;
+use crate::builder::ValueRange;
 
 /// Clearable builder value
 #[derive(Copy, Clone, Debug, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -44,6 +48,51 @@ pub trait IntoResettable<T> {
     fn into_resettable(self) -> Resettable<T>;
 }
 
+impl IntoResettable<char> for Option<char> {
+    fn into_resettable(self) -> Resettable<char> {
+        match self {
+            Some(s) => Resettable::Value(s),
+            None => Resettable::Reset,
+        }
+    }
+}
+
+impl IntoResettable<usize> for Option<usize> {
+    fn into_resettable(self) -> Resettable<usize> {
+        match self {
+            Some(s) => Resettable::Value(s),
+            None => Resettable::Reset,
+        }
+    }
+}
+
+impl IntoResettable<ArgAction> for Option<ArgAction> {
+    fn into_resettable(self) -> Resettable<ArgAction> {
+        match self {
+            Some(s) => Resettable::Value(s),
+            None => Resettable::Reset,
+        }
+    }
+}
+
+impl IntoResettable<ValueHint> for Option<ValueHint> {
+    fn into_resettable(self) -> Resettable<ValueHint> {
+        match self {
+            Some(s) => Resettable::Value(s),
+            None => Resettable::Reset,
+        }
+    }
+}
+
+impl IntoResettable<ValueParser> for Option<ValueParser> {
+    fn into_resettable(self) -> Resettable<ValueParser> {
+        match self {
+            Some(s) => Resettable::Value(s),
+            None => Resettable::Reset,
+        }
+    }
+}
+
 impl IntoResettable<StyledStr> for Option<&'static str> {
     fn into_resettable(self) -> Resettable<StyledStr> {
         match self {
@@ -74,6 +123,42 @@ impl IntoResettable<Str> for Option<&'static str> {
 impl<T> IntoResettable<T> for Resettable<T> {
     fn into_resettable(self) -> Resettable<T> {
         self
+    }
+}
+
+impl IntoResettable<char> for char {
+    fn into_resettable(self) -> Resettable<char> {
+        Resettable::Value(self)
+    }
+}
+
+impl IntoResettable<usize> for usize {
+    fn into_resettable(self) -> Resettable<usize> {
+        Resettable::Value(self)
+    }
+}
+
+impl IntoResettable<ArgAction> for ArgAction {
+    fn into_resettable(self) -> Resettable<ArgAction> {
+        Resettable::Value(self)
+    }
+}
+
+impl IntoResettable<ValueHint> for ValueHint {
+    fn into_resettable(self) -> Resettable<ValueHint> {
+        Resettable::Value(self)
+    }
+}
+
+impl<I: Into<ValueRange>> IntoResettable<ValueRange> for I {
+    fn into_resettable(self) -> Resettable<ValueRange> {
+        Resettable::Value(self.into())
+    }
+}
+
+impl<I: Into<ValueParser>> IntoResettable<ValueParser> for I {
+    fn into_resettable(self) -> Resettable<ValueParser> {
+        Resettable::Value(self.into())
     }
 }